11_Web Research Summarizer Agent/11. Web_Research Summarizer Agent/requirements.txt
openai
gradio
requests
beautifulsoup4
python-dotenv
pyyaml
duckduckgo-search
